[Misc] Resolve comments
