+++ examples/online_serving/disaggregated_encoder/README.md
@@ -1,0 +1,13 @@
+# Disaggregated Encoder
+
+This example contains scripts that demonstrate the disaggregated encoder (EPD) features of vLLM.
+
+Please refer to [Disaggregated Encoder Feature](../../../docs/features/disagg_encoder.md) for the detailed explanation for the EPD features.
+
+## Files
+
+- `disagg_epd_proxy.py` - Proxy to demonstrates XeYpZd (X encode instances, Y prefill instances, Z decode instances); Currently stable for 1e1p1d.
+- `disagg_1e1p1d_example.sh` - Setup 1e1p1d and run VisionArena benchmark.
+- `disagg_1e1pd_example.sh` - Setup 1e1pd and run VisionArena benchmark.
+
+Detailed explanations are commnented in the scripts.
